Revert "mako: touch: add error checking for reading descriptor"
Due to the side effect of the patch, firmware update might not be
recommenced when the firmware's in bad state. The firmware could be
in bad state when the firmware update's started in recovery mode and
system reboots before it's finished.

The error checking was added to detect I2C error after firmware update
but I2C error doesn't actually happen because of
e900ac0de811f62612480aa38fc3ac2d13fa2055, so revert it.

b/13532017

This reverts commit ca1e3b0e5f134aa837c8b808779eb2426ff96944.

Change-Id: I8916f7ca5ab9fe8a4685334349a882b580c26179
Signed-off-by: Jongrak Kwon <jongrak.kwon@lge.com>